Fix so that only one of the "testfaces" are wrong
The problem is RotateColors Clockwise(int nr). When the nr is 1 the code runs twice. When it should only run once. This is apperently not a problem for when the nr is 3. So far I dont know what happends when the nr is 2 but probably that will work to. The REAL issiue is that the loop runs once when nr is 0 and J is 0.